+++ algonaut_core/src/lib.rs
@@ -380,46 +380,6 @@ pub enum LogicSignature {
     DelegatedMultiSig(MultisigSignature),
 }
 
-#[derive(Debug)]
-pub struct ApiSignedLogic {
-    pub logic: Vec<u8>,
-    pub sig: Option<Signature>,
-    pub msig: Option<MultisigSignature>,
-    pub args: Vec<Vec<u8>>,
-}
-
-impl From<SignedLogic> for ApiSignedLogic {
-    fn from(s: SignedLogic) -> Self {
-        let (sig, msig) = match s.sig {
-            LogicSignature::ContractAccount => (None, None),
-            LogicSignature::DelegatedSig(sig) => (Some(sig), None),
-            LogicSignature::DelegatedMultiSig(msig) => (None, Some(msig)),
-        };
-        ApiSignedLogic {
-            logic: s.logic.bytes,
-            sig,
-            msig,
-            args: s.args,
-        }
-    }
-}
-
-impl Serialize for ApiSignedLogic {
-    fn serialize<S>(&self, serializer: S) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
-    where
-        S: Serializer,
-    {
-        // For some reason SerializeStruct ends up serializing as an array, so this explicitly serializes as a map
-        use serde::ser::SerializeMap;
-        let mut state = serializer.serialize_map(Some(4))?;
-        state.serialize_entry("l", &self.logic)?;
-        state.serialize_entry("arg", &self.args)?;
-        state.serialize_entry("sig", &self.sig)?;
-        state.serialize_entry("msig", &self.msig)?;
-        state.end()
-    }
-}
-
 pub trait ToMsgPack: Serialize {
     fn to_msg_pack(&self) -> Result<Vec<u8>, rmp_serde::encode::Error> {
         rmp_serde::to_vec_named(&self)
