+++ algonaut_transaction/src/lib.rs
@@ -1,12 +1,11 @@
 pub mod account;
-pub mod api_transaction;
+mod api_models;
 pub mod auction;
 pub mod builder;
 pub mod error;
 pub mod transaction;
 pub mod tx_group;
 
-pub use api_transaction::ApiSignedTransaction;
 pub use builder::{
     AcceptAsset, CallApplication, ClawbackAsset, ConfigureAsset, FreezeAsset, Pay, RegisterKey,
     TransferAsset, Txn,
