+++ examples/sign_submit_transaction.rs
@@ -1,6 +1,6 @@
 use algonaut::core::MicroAlgos;
 use algonaut::transaction::{Pay, Txn};
-use algonaut::{Algod, Kmd};
+use algonaut_client::{Algod, Kmd};
 use dotenv::dotenv;
 use std::env;
 use std::error::Error;
