+++ examples/wallet_restore.rs
@@ -1,6 +1,6 @@
 use algonaut::crypto::mnemonic;
 use algonaut::crypto::MasterDerivationKey;
-use algonaut::Kmd;
+use algonaut_client::Kmd;
 use dotenv::dotenv;
 use std::env;
 use std::error::Error;
