+++ src/lib.rs
@@ -17,9 +17,6 @@
 
 // Re-exports
 pub use algonaut_client as client;
-pub use algonaut_client::algod::Algod;
-pub use algonaut_client::indexer::Indexer;
-pub use algonaut_client::kmd::Kmd;
 pub use algonaut_core as core;
 pub use algonaut_crypto as crypto;
 pub use algonaut_transaction as transaction;
